Fix issue where downloading files over 2.1gb would cause the progress value to overflow

+++ src/main/java/com/dropbox/core/util/ProgressOutputStream.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.io.OutputStream;
 
 public class ProgressOutputStream extends OutputStream {
-    private int completed;
+    private long completed;
     private OutputStream underlying;
     private IOUtil.ProgressListener listener;
